requirements.txt
matplotlib==3.6.2
nltk==3.7
numpy==1.23.5
pandoc==2.3
Pillow==8.3.2
python-dateutil==2.8.2
regex==2022.10.31
textblob==0.17.1
tqdm==4.64.1
tzdata==2022.1
urllib3==1.26.9
whitenoise==5.3.0
wordcloud==1.8.2.2
pandas
seaborn
scikit-learn
textblob

